correction on angles
git-svn-id: file:///home/roberto/Git/svn@10971 1a721e8d-d38c-4e91-ab98-6a0be4d754af

Former-commit-id: a10adf8b7a5a73aa92c31ed7001384284e3a033a

+++ trunk/xmipp/java/src/xmipp/jni/TiltPairAligner.java
@@ -15,7 +15,7 @@ public class TiltPairAligner
 
 	public native Particle getTiltedParticle(int x1, int y1);
 	
-	public native int[] computeAngles();
+	public native double[] computeAngles();
 
 	// caching some ids
 	private static native void storeIds();
