+++ trunk/xmipp/java/src/xmipp/particlepicker/tiltpair/gui/TiltPairPickerJFrame.java
@@ -361,7 +361,7 @@ public class TiltPairPickerJFrame extends ParticlePickerJFrame
 	{
 		if(untiltedmic.getAddedCount() < 4)
 			return Math.PI/2;
-		int [] alphas = untiltedmic.getAngles();
+		double [] alphas = untiltedmic.getAngles();
 		double alpha = Math.toRadians(alphas[0]);
 		return alpha;
 	}
@@ -370,7 +370,7 @@ public class TiltPairPickerJFrame extends ParticlePickerJFrame
 	{
 		if(untiltedmic.getAddedCount() < 4)
 			return  Math.PI/2;
-		int [] alphas = untiltedmic.getAngles();
+		double [] alphas = untiltedmic.getAngles();
 		double alpha = Math.toRadians(alphas[1]);
 		return alpha;
 	}
